Add initial database migration and remove RedirectToError component
This commit introduces the initial database migration adding, among others, Groups, Services, and Details tables to the database. In addition, it removes the RedirectToError.razor component from the project. Moreover, the ServiceProduct model has been updated to include a new foreign key reference to the ProductGroup model and a new ConfigDbContext for handling the application's data has been added.