Refactor UI: Add Pricing page, remove Counter, update Services and Home
This commit includes the removal of the Counter page, addition of a new page - Pricing, and changes to the Services and Home page. The refactor also includes the change of theme settings in the MainLayout to Dark Mode. This aims to improve the user interface and functionality of the website.