Update render modes and implement contact form
The rendering mode of the 'Pricing' page was switched from 'InteractiveAuto' to 'InteractiveWebAssembly'. 'Contact' page was also modified where the existing contact form functionality gets a revamp, especially the error fields, and a new submission function was added. Additionally, a form with editable fields for user contact information was added in the front-end