Update fonts and remove carousel captions
The fonts used in the client settings have been updated to use "Forum" and "Cinzel" and the render mode for the About page has been set to InteractiveServer. Certain elements in the Carousel component have been removed to provide a cleaner user interface. A link to the new fonts has also been added in the App component.