Refactor and streamline code for contact form processing
This commit includes changes made to various files in order to streamline the processing of the contact form. This involves changes to input field widths and methods related to form submission in the front-end Razor Pages. In addition, the data mapping configuration in AutoMapper for ClientContactDTO has been updated. The code in the back-end has been further refactored for improved readability and performance. This includes removal of redundant code from Contact.razor.cs and updating ToString() methods in DTO classes for better data representation.